+++ cognite/replicator/datapoints.py
@@ -1,5 +1,6 @@
 import logging
 import multiprocessing as mp
+import re
 from datetime import datetime
 from math import ceil, floor
 from typing import Any, Callable, List, Optional, Tuple, Union
@@ -225,6 +226,7 @@ def replicate(
     timerange_transform: Optional[Callable[[Tuple[int, int]], Tuple[int, int]]] = None,
     start: Union[int, str] = None,
     end: Union[int, str] = None,
+    exclude_pattern: str = None,
 ):
     """
     Replicates data points from the source project into the destination project for all time series that
@@ -243,19 +245,38 @@ def replicate(
         timerange_transform: Function to set the time range boundaries (start, end) arbitrarily.
         start: Timestamp to start replication onwards from; if not specified starts at most recent datapoint
         end: If specified, limits replication to datapoints earlier than the end time
+        exclude_pattern: Regex pattern; time series whose names match will not be replicated from
     """
-    if external_ids is not None:
+
+    if external_ids and exclude_pattern:
+        raise ValueError(
+            f"List of time series AND a regex exclusion rule was given! Either remove the filter {exclude_pattern} or the list of time series {external_ids}"
+        )
+    elif external_ids is not None:  # Specified list of time series is given
         ts_src = client_src.time_series.retrieve_multiple(external_ids=external_ids)
         ts_dst = client_dst.time_series.retrieve_multiple(external_ids=external_ids)
+        src_ext_id_list = [ts_obj.external_id for ts_obj in ts_src]
     else:
         ts_src = client_src.time_series.list(limit=None)
         ts_dst = client_dst.time_series.list(limit=None)
-
+        filtered_ts_src = []
+        skipped_ts = []
+        if exclude_pattern:  # Filtering based on regex rule given
+            compiled_re = re.compile(exclude_pattern)
+            for ts in ts_src:
+                if compiled_re.search(ts.external_id):
+                    skipped_ts.append(ts.external_id)
+                elif not compiled_re.search(ts.external_id):
+                    filtered_ts_src.append(ts.external_id)
+            src_ext_id_list = filtered_ts_src
+            logging.info(f"Excluding time series: {skipped_ts}, due to regex rule: {exclude_pattern}")
+            # Should probably change to logging.debug after a while
+        else:  # Expects to replicate all shared time series
+            src_ext_id_list = [ts_obj.external_id for ts_obj in ts_src]
     logging.info(f"Number of time series in source: {len(ts_src)}")
     logging.info(f"Number of time series in destination: {len(ts_dst)}")
 
-    src_ext_id_list = [ts_id.external_id for ts_id in ts_src]
-    dst_ext_id_list = set([ts_id.external_id for ts_id in ts_dst])
+    dst_ext_id_list = set([ts_obj.external_id for ts_obj in ts_dst])
     shared_external_ids = [ext_id for ext_id in src_ext_id_list if ext_id in dst_ext_id_list and ext_id]
     logging.info(
         f"Number of common time series external ids between destination and source: {len(shared_external_ids)}"